unified tests to client-wrapper-spec
