--- packages/elasticsearch-store/test/client-wrapper-functions/exists-spec.ts
@@ -1,81 +1,0 @@
-import { debugLogger, DataEntity } from '@terascope/utils';
-import { createClient, WrappedClient, Semver, } from '../../src';
-import {
-    upload,
-    cleanupIndex,
-    waitForData,
-    getDistributionAndVersion
-} from '../helpers/elasticsearch';
-import { data } from '../helpers/data';
-
-const testLogger = debugLogger('create-client-test');
-
-const {
-    version,
-    host,
-    distribution
-} = getDistributionAndVersion();
-
-const semver = version.split('.').map((i) => parseInt(i, 10)) as Semver;
-
-describe('exists', () => {
-    let wrappedClient: WrappedClient;
-    const index = 'test-exists';
-    const docType = '_doc';
-    let client: any;
-
-    beforeAll(async () => {
-        ({ client } = await createClient({ node: host }, testLogger));
-
-        wrappedClient = new WrappedClient(client, distribution, semver);
-
-        const testData = data.slice(0, 1).map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
-
-        await cleanupIndex(client, index);
-
-        await upload(client, { index, type: docType }, testData);
-        await waitForData(client, index, 1);
-    });
-
-    afterAll(async () => {
-        await cleanupIndex(client, index);
-    });
-
-    it('should return true if the record exists', async () => {
-        const params = {
-            id: '1',
-            index
-        };
-
-        const resp = await wrappedClient.exists(params);
-
-        expect(resp).toBe(true);
-    });
-
-    it('should return false if the record does not exist', async () => {
-        const params = {
-            id: '2',
-            index
-        };
-
-        const resp = await wrappedClient.exists(params);
-
-        expect(resp).toBe(false);
-    });
-
-    it('should return response with complex params', async () => {
-        const params = {
-            id: '1',
-            index,
-            type: '_doc',
-            preference: '_local',
-            realtime: true,
-            refresh: false,
-            routing: '1'
-        };
-
-        const resp = await wrappedClient.exists(params);
-
-        expect(resp).toBe(true);
-    });
-});
