+++ packages/elasticsearch-store/test/client-wrapper-spec.ts
@@ -1,19 +1,19 @@
+/* eslint-disable jest/no-disabled-tests */
 import { DataEntity, debugLogger, toNumber } from '@terascope/utils';
-import { ElasticsearchDistribution } from '@terascope/types';
 import {
-    createClient, WrappedClient, Semver,
-    ClientParams
+    createClient,
+    WrappedClient,
+    Semver,
+    ClientParams,
 } from '../src';
 import {
-    upload, cleanupIndex, waitForData,
-    formatUploadData
+    upload,
+    cleanupIndex,
+    waitForData,
+    formatUploadData,
+    getDistributionAndVersion,
+    getTotalFormat
 } from './helpers/elasticsearch';
-import {
-    ELASTICSEARCH_HOST,
-    ELASTICSEARCH_VERSION,
-    OPENSEARCH_HOST,
-    OPENSEARCH_VERSION
-} from './helpers/config';
 import { data } from './helpers/data';
 
 describe('can create an elasticsearch or opensearch client', () => {
@@ -21,19 +21,18 @@ describe('can create an elasticsearch or opensearch client', () => {
     const docType = '_doc';
 
     const testLogger = debugLogger('create-client-test');
-    const config = { node: '' };
-    let expectedDistribution: ElasticsearchDistribution;
-    let expectedVersion: string;
 
-    if (process.env.TEST_OPENSEARCH != null) {
-        config.node = OPENSEARCH_HOST;
-        expectedDistribution = ElasticsearchDistribution.opensearch;
-        expectedVersion = OPENSEARCH_VERSION;
-    } else {
-        config.node = ELASTICSEARCH_HOST;
-        expectedDistribution = ElasticsearchDistribution.elasticsearch;
-        expectedVersion = ELASTICSEARCH_VERSION;
-    }
+    const {
+        version,
+        host,
+        distribution
+    } = getDistributionAndVersion();
+
+    const config = { node: host };
+
+    const semver = version.split('.').map(toNumber) as unknown as Semver;
+
+    const majorVersion = semver[0];
 
     let wrappedClient: WrappedClient;
     let client: any;
@@ -43,15 +42,55 @@ describe('can create an elasticsearch or opensearch client', () => {
 
         await cleanupIndex(client, index);
 
-        const semver = expectedVersion.split('.').map(toNumber) as unknown as Semver;
         if (semver.length !== 3) {
-            throw new Error(`Expected version to follow semver format (major.minor.patch) got ${expectedVersion}`);
+            throw new Error(`Expected version to follow semver format (major.minor.patch) got ${version}`);
         }
 
         await upload(client, { index, type: docType }, data);
         await waitForData(client, index, 1000);
 
-        wrappedClient = new WrappedClient(client, expectedDistribution, semver);
+        wrappedClient = new WrappedClient(client, distribution, semver);
+    });
+
+    describe('info', () => {
+        it('should return info about the cluster', async () => {
+            const resp = await wrappedClient.info();
+
+            if (distribution === 'elasticsearch') {
+                if (version.split('.')[0] === '8') {
+                    expect(resp.cluster_name).toBe('docker-cluster');
+                } else {
+                    expect(resp.cluster_name).toBe(distribution);
+                }
+            }
+
+            if (distribution === 'opensearch') {
+                expect(resp.cluster_name).toBe('docker-cluster');
+                expect(resp.version.distribution).toBe('opensearch');
+            }
+
+            expect(resp.version.number).toBe(version);
+        });
+
+        it('should throw an error if tried on a non-supported distribution', async () => {
+            const badDistribution = new WrappedClient(client, distribution, [3, 2, 1]);
+
+            await expect(() => badDistribution.ping()).rejects.toThrowError(`Unsupported ${distribution} version: 3.2.1`);
+        });
+    });
+
+    describe('ping', () => {
+        it('should return true if cluster is running', async () => {
+            const resp = await wrappedClient.ping();
+
+            expect(resp).toBe(true);
+        });
+
+        it('should throw an error if tried on a non-supported distribution', async () => {
+            const badDistribution = new WrappedClient(client, distribution, [10, 0, 0]);
+
+            await expect(() => badDistribution.ping()).rejects.toThrowError(`Unsupported ${distribution} version: 10.0.0`);
+        });
     });
 
     describe('count', () => {
@@ -190,18 +229,23 @@ describe('can create an elasticsearch or opensearch client', () => {
 
     describe('update', () => {
         const testIndex = `${index}_update_method`;
+
         const doc = {
             some: 'newThing',
             bool: true,
             obj: { other: 'thing' },
             method: 'update'
         };
-        const id = '123412341234';
+        const id1 = '123412341234';
+        const id2 = '123412341235';
 
-        const record = DataEntity.make(doc, { _key: id });
+        const records = [
+            DataEntity.make(doc, { _key: id1 }),
+            DataEntity.make(doc, { _key: id2 }),
+        ];
 
         beforeAll(async () => {
-            const testData = formatUploadData(testIndex, [record]);
+            const testData = formatUploadData(testIndex, records);
             await upload(client, { index: testIndex }, testData);
         });
 
@@ -215,7 +259,7 @@ describe('can create an elasticsearch or opensearch client', () => {
             };
 
             const query: ClientParams.UpdateParams = {
-                id,
+                id: id1,
                 index: testIndex,
                 type: docType,
                 body: { doc: updatedDoc },
@@ -225,7 +269,30 @@ describe('can create an elasticsearch or opensearch client', () => {
             const response = await wrappedClient.update(query);
 
             expect(response).toHaveProperty('_index', testIndex);
-            expect(response).toHaveProperty('_id', id);
+            expect(response).toHaveProperty('_id', id1);
+            expect(response).toHaveProperty('_version', 2);
+            expect(response).toHaveProperty('result', 'updated');
+            expect(response).toHaveProperty('forced_refresh', true);
+            expect(response).toHaveProperty('_shards');
+            expect(response).toHaveProperty('_seq_no');
+        });
+
+        it('can update a new record without a type', async () => {
+            const updatedDoc = {
+                iLoveUpdates: true
+            };
+
+            const query: ClientParams.UpdateParams = {
+                id: id2,
+                index: testIndex,
+                body: { doc: updatedDoc },
+                refresh: true
+            };
+
+            const response = await wrappedClient.update(query);
+
+            expect(response).toHaveProperty('_index', testIndex);
+            expect(response).toHaveProperty('_id', id2);
             expect(response).toHaveProperty('_version', 2);
             expect(response).toHaveProperty('result', 'updated');
             expect(response).toHaveProperty('forced_refresh', true);
@@ -354,6 +421,470 @@ describe('can create an elasticsearch or opensearch client', () => {
         });
     });
 
+    describe('mget', () => {
+        const mgetIndex = 'test-mget';
+
+        beforeAll(async () => {
+            const testData = data.slice(0, 10)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await cleanupIndex(client, mgetIndex);
+
+            await upload(client, { index: mgetIndex, type: docType }, testData);
+            await waitForData(client, mgetIndex, 10);
+        });
+
+        afterAll(async () => {
+            await cleanupIndex(client, mgetIndex);
+        });
+
+        it('should handle docs in body property', async () => {
+            const params = {
+                body: {
+                    docs:
+                        [
+                            { _index: 'test-mget', _id: '1' },
+                            { _index: 'test-mget', _id: '7' },
+                            { _index: 'test-mget', _id: '4' }
+                        ]
+                }
+            };
+
+            const resp = await wrappedClient.mget(params) as any;
+
+            expect(resp.docs[0]._source?.uuid).toBe('b23a8550-0081-453f-9e80-93a90782a5bd');
+            expect(resp.docs[1]._source?.uuid).toBe('b284b6c9-43bb-4c59-a4e4-fdb17b004300');
+            expect(resp.docs[2]._source?.uuid).toBe('bd920141-45b3-41fd-8eea-b1640a2fa3d2');
+        });
+
+        it('should handle docs with _type doc objects', async () => {
+            const params = {
+                body: {
+                    docs:
+                        [
+                            { _index: 'test-mget', _id: '1', _type: docType },
+                            { _index: 'test-mget', _id: '7', _type: docType },
+                            { _index: 'test-mget', _id: '4', _type: docType }
+                        ]
+                }
+            };
+
+            const resp = await wrappedClient.mget(params) as any;
+
+            expect(resp.docs[0]._source?.uuid).toBe('b23a8550-0081-453f-9e80-93a90782a5bd');
+            expect(resp.docs[1]._source?.uuid).toBe('b284b6c9-43bb-4c59-a4e4-fdb17b004300');
+            expect(resp.docs[2]._source?.uuid).toBe('bd920141-45b3-41fd-8eea-b1640a2fa3d2');
+        });
+
+        it('should handle ids in body property', async () => {
+            const params = {
+                index: mgetIndex,
+                body: {
+                    ids: ['1', '7', '4']
+                }
+            };
+
+            const resp = await wrappedClient.mget(params) as any;
+
+            expect(resp.docs.length).toBe(3);
+            expect(resp.docs[0]._source?.uuid).toBe('b23a8550-0081-453f-9e80-93a90782a5bd');
+            expect(resp.docs[1]._source?.uuid).toBe('b284b6c9-43bb-4c59-a4e4-fdb17b004300');
+            expect(resp.docs[2]._source?.uuid).toBe('bd920141-45b3-41fd-8eea-b1640a2fa3d2');
+        });
+
+        it('should handle ids in body property with type', async () => {
+            const params = {
+                index: mgetIndex,
+                type: docType,
+                body: {
+                    ids: ['1', '7', '4']
+                }
+            };
+
+            const resp = await wrappedClient.mget(params) as any;
+
+            expect(resp.docs.length).toBe(3);
+            expect(resp.docs[0]._source?.uuid).toBe('b23a8550-0081-453f-9e80-93a90782a5bd');
+            expect(resp.docs[1]._source?.uuid).toBe('b284b6c9-43bb-4c59-a4e4-fdb17b004300');
+            expect(resp.docs[2]._source?.uuid).toBe('bd920141-45b3-41fd-8eea-b1640a2fa3d2');
+        });
+    });
+
+    describe('search', () => {
+        const searchIndex = 'test-search';
+
+        const total = getTotalFormat(distribution, semver[0], 1);
+
+        beforeAll(async () => {
+            const testData = data.slice(0, 10)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await cleanupIndex(client, searchIndex);
+
+            await upload(client, { index: searchIndex, type: docType }, testData);
+            await waitForData(client, searchIndex, 10);
+        });
+
+        afterAll(async () => {
+            await cleanupIndex(client, searchIndex);
+        });
+
+        it('should return record on q search', async () => {
+            const params = {
+                index: searchIndex,
+                q: 'uuid:bea4086e-6f2e-4f4b-a1bf-c20330f92e8c'
+            };
+
+            const resp = await wrappedClient.search(params);
+
+            expect(resp.hits.total).toEqual(total);
+
+            expect(resp.hits.hits[0]._source).toEqual({
+                ip: '143.174.175.238',
+                userAgent: 'Mozilla/5.0 (Windows; U; Windows NT 6.3) AppleWebKit/531.0.0 (KHTML, like Gecko) Chrome/22.0.897.0 Safari/531.0.0',
+                url: 'http://dedrick.biz',
+                uuid: 'bea4086e-6f2e-4f4b-a1bf-c20330f92e8c',
+                created: '2019-04-26T15:00:23.213+00:00',
+                ipv6: '5b2a:9397:6e8c:ac74:63a0:799c:00b5:92d2',
+                location: '88.04393, -35.42878',
+                bytes: 188644
+            });
+        });
+
+        it('should return record with type and index in search', async () => {
+            const params = {
+                type: docType,
+                index: searchIndex,
+                q: 'uuid:bea4086e-6f2e-4f4b-a1bf-c20330f92e8c'
+            };
+
+            const resp = await wrappedClient.search(params);
+
+            expect(resp.hits.total).toEqual(total);
+            expect(resp.hits.hits[0]._source).toEqual({
+                ip: '143.174.175.238',
+                userAgent: 'Mozilla/5.0 (Windows; U; Windows NT 6.3) AppleWebKit/531.0.0 (KHTML, like Gecko) Chrome/22.0.897.0 Safari/531.0.0',
+                url: 'http://dedrick.biz',
+                uuid: 'bea4086e-6f2e-4f4b-a1bf-c20330f92e8c',
+                created: '2019-04-26T15:00:23.213+00:00',
+                ipv6: '5b2a:9397:6e8c:ac74:63a0:799c:00b5:92d2',
+                location: '88.04393, -35.42878',
+                bytes: 188644
+            });
+        });
+
+        it('should return record with body search', async () => {
+            const params = {
+                type: docType,
+                index: searchIndex,
+                body: {
+                    query: {
+                        match: {
+                            uuid: 'bea4086e-6f2e-4f4b-a1bf-c20330f92e8c'
+                        }
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.search(params);
+
+            expect(resp.hits.total).toEqual(total);
+            expect(resp.hits.hits[0]._source).toEqual({
+                ip: '143.174.175.238',
+                userAgent: 'Mozilla/5.0 (Windows; U; Windows NT 6.3) AppleWebKit/531.0.0 (KHTML, like Gecko) Chrome/22.0.897.0 Safari/531.0.0',
+                url: 'http://dedrick.biz',
+                uuid: 'bea4086e-6f2e-4f4b-a1bf-c20330f92e8c',
+                created: '2019-04-26T15:00:23.213+00:00',
+                ipv6: '5b2a:9397:6e8c:ac74:63a0:799c:00b5:92d2',
+                location: '88.04393, -35.42878',
+                bytes: 188644
+            });
+        });
+    });
+
+    describe('msearch', () => {
+        const msearchIndex = 'test-msearch';
+
+        const total = getTotalFormat(distribution, semver[0], 1);
+
+        beforeAll(async () => {
+            const testData = data.slice(0, 10)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await cleanupIndex(client, msearchIndex);
+
+            await upload(client, { index: msearchIndex, type: docType }, testData);
+            await waitForData(client, msearchIndex, 10);
+        });
+
+        afterAll(async () => {
+            await cleanupIndex(client, msearchIndex);
+        });
+
+        it('should return requested records', async () => {
+            const params = {
+                body: [
+                    { index: msearchIndex },
+                    { query: { match: { uuid: 'bd920141-45b3-41fd-8eea-b1640a2fa3d2' } } },
+                    { index: msearchIndex },
+                    { query: { match: { uuid: 'b23a8550-0081-453f-9e80-93a90782a5bd' } } }
+                ]
+            };
+
+            const resp = await wrappedClient.msearch(params);
+
+            expect(resp.responses.length).toBe(2);
+
+            expect(resp.responses[0].hits.total).toEqual(total);
+            expect(resp.responses[0].hits.hits[0]._source?.uuid).toBe('bd920141-45b3-41fd-8eea-b1640a2fa3d2');
+            expect(resp.responses[1].hits.total).toEqual(total);
+            expect(resp.responses[1].hits.hits[0]._source?.uuid).toBe('b23a8550-0081-453f-9e80-93a90782a5bd');
+        });
+
+        it('should handle type in params and return requested records', async () => {
+            const params = {
+                index: msearchIndex,
+                type: docType,
+                body: [
+                    { index: msearchIndex, type: docType },
+                    { query: { match: { uuid: 'bd920141-45b3-41fd-8eea-b1640a2fa3d2' } } },
+                    { index: msearchIndex, type: docType },
+                    { query: { match: { uuid: 'b23a8550-0081-453f-9e80-93a90782a5bd' } } }
+                ]
+            };
+
+            const resp = await wrappedClient.msearch(params);
+
+            expect(resp.responses.length).toBe(2);
+
+            expect(resp.responses[0].hits.total).toEqual(total);
+            expect(resp.responses[0].hits.hits[0]._source?.uuid).toBe('bd920141-45b3-41fd-8eea-b1640a2fa3d2');
+            expect(resp.responses[1].hits.total).toEqual(total);
+            expect(resp.responses[1].hits.hits[0]._source?.uuid).toBe('b23a8550-0081-453f-9e80-93a90782a5bd');
+        });
+
+        it('should handle ccs_minimize_roundtrips in params and return requested records', async () => {
+            const params = {
+                ccs_minimize_roundtrips: true,
+                body: [
+                    { index: msearchIndex, type: docType },
+                    { query: { match: { uuid: 'bd920141-45b3-41fd-8eea-b1640a2fa3d2' } } },
+                    { index: msearchIndex, type: docType },
+                    { query: { match: { uuid: 'b23a8550-0081-453f-9e80-93a90782a5bd' } } }
+                ]
+            };
+
+            const resp = await wrappedClient.msearch(params);
+
+            expect(resp.responses.length).toBe(2);
+
+            expect(resp.responses[0].hits.total).toEqual(total);
+            expect(resp.responses[0].hits.hits[0]._source?.uuid).toBe('bd920141-45b3-41fd-8eea-b1640a2fa3d2');
+            expect(resp.responses[1].hits.total).toEqual(total);
+            expect(resp.responses[1].hits.hits[0]._source?.uuid).toBe('b23a8550-0081-453f-9e80-93a90782a5bd');
+        });
+    });
+
+    describe('exists', () => {
+        const existsIndex = 'test-exists';
+
+        beforeAll(async () => {
+            const testData = data.slice(0, 1)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await cleanupIndex(client, existsIndex);
+
+            await upload(client, { index: existsIndex, type: docType }, testData);
+            await waitForData(client, existsIndex, 1);
+        });
+
+        afterAll(async () => {
+            await cleanupIndex(client, existsIndex);
+        });
+
+        it('should return true if the record exists', async () => {
+            const params = {
+                id: '1',
+                index: existsIndex
+            };
+
+            const resp = await wrappedClient.exists(params);
+
+            expect(resp).toBe(true);
+        });
+
+        it('should return false if the record does not exist', async () => {
+            const params = {
+                id: '2',
+                index: existsIndex
+            };
+
+            const resp = await wrappedClient.exists(params);
+
+            expect(resp).toBe(false);
+        });
+
+        it('should return response with complex params', async () => {
+            const params = {
+                id: '1',
+                index: existsIndex,
+                type: '_doc',
+                preference: '_local',
+                realtime: true,
+                refresh: false,
+                routing: '1'
+            };
+
+            const resp = await wrappedClient.exists(params);
+
+            expect(resp).toBe(true);
+        });
+    });
+
+    describe('reindex', () => {
+        const reindexIndex = 'test-reindex';
+
+        const testIndices = [
+            'test-reindex2',
+            'test-reindex3',
+            'test-reindex4',
+            'test-reindex5',
+            'test-reindex6'
+        ];
+
+        beforeAll(async () => {
+            const testData = data.slice(0, 10)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await cleanupIndex(client, reindexIndex);
+
+            await upload(client, { index: reindexIndex, type: docType }, testData);
+            await waitForData(client, reindexIndex, 10);
+        });
+
+        afterAll(async () => {
+            await cleanupIndex(client, reindexIndex);
+            await Promise.all(testIndices.map((i) => cleanupIndex(client, i)));
+        });
+
+        it('should reindex test index', async () => {
+            const params = {
+                body: {
+                    source: {
+                        index: reindexIndex
+                    },
+                    dest: {
+                        index: 'test-reindex2'
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.reindex(params);
+
+            expect(resp.total).toBe(10);
+            expect(resp.created).toBe(10);
+            expect(resp.failures.length).toBe(0);
+        });
+
+        it('should handle types in reindex params', async () => {
+            const params = {
+                body: {
+                    source: {
+                        index: reindexIndex
+                    },
+                    dest: {
+                        index: 'test-reindex3',
+                        type: '_doc'
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.reindex(params);
+
+            expect(resp.total).toBe(10);
+            expect(resp.created).toBe(10);
+            expect(resp.failures.length).toBe(0);
+        });
+
+        it('should handle a script in reindex params', async () => {
+            const params: ClientParams.ReindexParams = {
+                body: {
+                    source: {
+                        index: reindexIndex
+                    },
+                    dest: {
+                        index: 'test-reindex4',
+                        type: '_doc'
+                    },
+                    script: {
+                        source: 'ctx._source.host = ctx._source.remove("url")',
+                        lang: 'painless'
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.reindex(params);
+
+            expect(resp.total).toBe(10);
+            expect(resp.created).toBe(10);
+            expect(resp.failures.length).toBe(0);
+        });
+
+        it('should handle a query in source params', async () => {
+            const params = {
+                body: {
+                    source: {
+                        index: reindexIndex,
+                        query: { match: { uuid: 'bd920141-45b3-41fd-8eea-b1640a2fa3d2' } }
+                    },
+                    dest: {
+                        index: 'test-reindex6',
+                        type: '_doc'
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.reindex(params);
+
+            expect(resp.total).toBe(1);
+            expect(resp.created).toBe(1);
+            expect(resp.failures.length).toBe(0);
+        });
+
+        it('should handle additional reindex params', async () => {
+            const params: ClientParams.ReindexParams = {
+                refresh: true,
+                max_docs: 5,
+                wait_for_active_shards: 1,
+                requests_per_second: 100,
+                body: {
+                    conflicts: 'proceed',
+                    source: {
+                        index: reindexIndex,
+                        _source: ['url', 'ip', 'created', 'userAgent']
+                    },
+                    dest: {
+                        index: 'test-reindex5',
+                        type: '_doc',
+                        op_type: 'create'
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.reindex(params);
+
+            if (distribution === 'elasticsearch') {
+                if (semver[0] === 6) {
+                    expect(resp.total).toBe(10);
+                    expect(resp.created).toBe(10);
+                } else {
+                    expect(resp.total).toBe(5);
+                    expect(resp.created).toBe(5);
+                }
+            }
+        });
+    });
+
     describe('cluster.getSettings', () => {
         it('can fetch settings from the cluster', async () => {
             const response = await wrappedClient.cluster.getSettings();
@@ -432,10 +963,576 @@ describe('can create an elasticsearch or opensearch client', () => {
             expect(node).toHaveProperty('ip');
             expect(node).toHaveProperty('roles');
             expect(node).toHaveProperty('indices');
-            expect(node).toHaveProperty('indices.docs.count', 1000);
+            expect(node).toHaveProperty('indices.docs.count');
             expect(node).toHaveProperty('os');
             expect(node).toHaveProperty('process');
             expect(node).toHaveProperty('jvm');
+        });
+    });
+
+    describe('indices.get', () => {
+        const testIndex = 'test-indices-get';
+
+        beforeAll(async () => {
+            const testData = data.slice(0, 1)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await cleanupIndex(client, testIndex);
+
+            await upload(client, { index: testIndex, type: docType }, testData);
+            await waitForData(client, testIndex, 1);
+        });
+
+        afterAll(async () => {
+            await cleanupIndex(client, testIndex);
+        });
+
+        it('should return mappings and settings for the index', async () => {
+            const params = {
+                index: testIndex
+            };
+
+            const resp = await wrappedClient.indices.get(params);
+
+            expect(resp).toHaveProperty(testIndex);
+            expect(resp[testIndex]).toHaveProperty('settings');
+            expect(resp[testIndex]).toHaveProperty('mappings');
+        });
+
+        it('should return mappings and settings for multiple indices', async () => {
+            const index2 = 'test-indices-get2';
+
+            const testData = data.slice(0, 1)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await upload(client, { index: index2, type: docType }, testData);
+
+            const params = {
+                index: 'test-indices-g*'
+            };
+
+            const resp = await wrappedClient.indices.get(params);
+
+            expect(resp).toHaveProperty(testIndex);
+            expect(resp[testIndex]).toHaveProperty('settings');
+            expect(resp[testIndex]).toHaveProperty('mappings');
+
+            expect(resp).toHaveProperty(index2);
+            expect(resp[index2]).toHaveProperty('settings');
+            expect(resp[index2]).toHaveProperty('mappings');
+        });
+
+        it('should handle detailed settings and mappings and settings for the index', async () => {
+            const params: ClientParams.IndicesGetParams = {
+                index: testIndex,
+                include_type_name: true,
+                local: true,
+                ignore_unavailable: true,
+                allow_no_indices: true,
+                expand_wildcards: 'none',
+                flat_settings: true,
+                include_defaults: true,
+                master_timeout: '60s'
+            };
+
+            const resp = await wrappedClient.indices.get(params);
+
+            expect(resp).toHaveProperty(testIndex);
+            expect(resp[testIndex]).toHaveProperty('settings');
+            expect(resp[testIndex]).toHaveProperty('mappings');
+        });
+    });
+
+    describe('indices.exists', () => {
+        const testIndex = 'test-indices-exists';
+        const anotherIndex = 'test-indices-exists2';
+
+        beforeAll(async () => {
+            const testData = data.slice(0, 1)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await cleanupIndex(client, testIndex);
+            await cleanupIndex(client, anotherIndex);
+
+            await upload(client, { index: testIndex, type: docType }, testData);
+            await upload(client, { index: anotherIndex, type: docType }, testData);
+            await waitForData(client, testIndex, 1);
+        });
+
+        afterAll(async () => {
+            await Promise.all([
+                cleanupIndex(client, testIndex),
+                cleanupIndex(client, anotherIndex)
+            ]);
+        });
+
+        it('should return true is index exists', async () => {
+            const params = {
+                index: testIndex
+            };
+
+            const resp = await wrappedClient.indices.exists(params);
+
+            expect(resp).toBeTrue();
+        });
+
+        it('should return true if index exists with parameters', async () => {
+            const params: ClientParams.IndicesExistsParams = {
+                index: testIndex,
+                allow_no_indices: false,
+                expand_wildcards: 'open',
+                flat_settings: true,
+                include_defaults: true,
+                ignore_unavailable: true,
+                local: true
+            };
+
+            const resp = await wrappedClient.indices.exists(params);
+
+            expect(resp).toBeTrue();
+        });
+
+        it('should return true if all indices exist', async () => {
+            const params: ClientParams.IndicesExistsParams = {
+                index: [testIndex, anotherIndex],
+                allow_no_indices: false,
+                expand_wildcards: 'open',
+                flat_settings: true,
+                include_defaults: true,
+                ignore_unavailable: true,
+                local: true
+            };
+
+            const resp = await wrappedClient.indices.exists(params);
+
+            expect(resp).toBeTrue();
+        });
+    });
+
+    describe('indices.create', () => {
+        const testIndex = 'test-indices-create';
+        const otherIndex = 'test-indices-create-with-settings-type';
+        const anotherIndex = 'test-indices-create-with-settings';
+
+        afterAll(async () => {
+            await Promise.all([
+                cleanupIndex(client, testIndex),
+                cleanupIndex(client, otherIndex),
+                cleanupIndex(client, anotherIndex),
+            ]);
+        });
+
+        it('should create new index', async () => {
+            const params = {
+                index: testIndex
+            };
+
+            const resp = await wrappedClient.indices.create(params);
+
+            expect(resp.acknowledged).toBeTrue();
+            expect(resp.shards_acknowledged).toBeTrue();
+            expect(resp.index).toBe(testIndex);
+        });
+
+        it('should create new index with mappings, alias, and settings if types present', async () => {
+            const params = {
+                index: otherIndex,
+                include_type_name: true,
+                body: {
+                    aliases: {
+                        'test-indices-yo-mama': {
+                            is_write_index: true,
+                            is_hidden: false
+                        }
+                    },
+                    mappings: {
+                        _doc: {
+                            properties: {
+                                _key: { type: 'keyword' },
+                                name: { type: 'text' },
+                                age: { type: 'short' }
+                            }
+                        }
+                    },
+                    settings: {
+                        number_of_shards: 3,
+                        number_of_replicas: 2,
+                        max_result_window: 10
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.indices.create(params);
+
+            expect(resp.acknowledged).toBeTrue();
+            expect(resp.shards_acknowledged).toBeTrue();
+            expect(resp.index).toBe(otherIndex);
+        });
+
+        it('should create new index with mappings, alias, and settings without types', async () => {
+            const params = {
+                index: anotherIndex,
+                body: {
+                    aliases: {
+                        'test-indices-yo-mama2': {
+                            is_write_index: true,
+                            is_hidden: false
+                        }
+                    },
+                    mappings: {
+                        properties: {
+                            _key: { type: 'keyword' },
+                            name: { type: 'text' },
+                            age: { type: 'short' }
+                        }
+                    },
+                    settings: {
+                        number_of_shards: 3,
+                        number_of_replicas: 2,
+                        max_result_window: 10
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.indices.create(params);
+
+            expect(resp.acknowledged).toBeTrue();
+            expect(resp.shards_acknowledged).toBeTrue();
+            expect(resp.index).toBe(anotherIndex);
+        });
+    });
+
+    describe('indices.delete', () => {
+        const testIndex = 'test-indices-delete';
+
+        beforeAll(async () => {
+            const testData = data.slice(0, 1)
+                .map((doc, i) => DataEntity.make(doc, { _key: i + 1 }));
+
+            await cleanupIndex(client, testIndex);
+
+            await upload(client, { index: testIndex, type: docType }, testData);
+            await waitForData(client, testIndex, 1);
+        });
+
+        afterAll(async () => {
+            await cleanupIndex(client, testIndex);
+        });
+
+        it('should delete the index', async () => {
+            const params = {
+                index: testIndex
+            };
+
+            const resp = await wrappedClient.indices.delete(params);
+
+            expect(resp.acknowledged).toBeTrue();
+        });
+    });
+
+    describe('indices.putTemplate', () => {
+        const tempName = 'test-template-put';
+
+        afterAll(async () => {
+            await client.indices.deleteTemplate({ name: tempName });
+        });
+
+        it('should save the template to the cluster', async () => {
+            const params: ClientParams.IndicesPutTemplateParams = {
+                name: tempName,
+                include_type_name: false,
+                order: 0,
+                create: true,
+                master_timeout: '60s',
+                body: {
+                    index_patterns: ['test-put-template'],
+                    settings: {
+                        number_of_shards: 3,
+                        number_of_replicas: 2
+                    },
+                    mappings: {
+                        properties: {
+                            name: { type: 'keyword' },
+                            uuid: { type: 'keyword' },
+                            created: { type: 'date' }
+                        }
+                    },
+                    aliases: {
+                        template_test: {}
+                    }
+                }
+            };
+
+            const resp = await wrappedClient.indices.putTemplate(params);
+
+            expect(resp.acknowledged).toBeTrue();
+        });
+    });
+
+    describe('indices.getTemplate', () => {
+        const tempName = 'test-template-get';
+
+        const settings = {
+            number_of_shards: '3',
+            number_of_replicas: '2'
+        };
+
+        const mappings = {
+            properties: {
+                name: { type: 'keyword' },
+                uuid: { type: 'keyword' },
+                created: { type: 'date' }
+            }
+        };
+
+        const indexPatterns = ['test-template-get*'];
+
+        const aliases = { template_test: {} };
+
+        beforeAll(async () => {
+            await wrappedClient.indices.putTemplate({
+                name: tempName,
+                body: {
+                    index_patterns: indexPatterns,
+                    settings,
+                    mappings,
+                    aliases
+                }
+            });
+        });
+
+        afterAll(async () => {
+            await client.indices.deleteTemplate({ name: tempName });
+        });
+
+        it('should return template', async () => {
+            const params: ClientParams.IndicesGetTemplateParams = {
+                name: tempName,
+                master_timeout: '60s',
+                local: false,
+                flat_settings: false
+            };
+
+            const resp = await wrappedClient.indices.getTemplate(params);
+
+            if (distribution === 'elasticsearch' && majorVersion === 6) {
+                if (semver[0] === 6) {
+                    expect(resp).toEqual({
+                        [tempName]: {
+                            order: 0,
+                            index_patterns: indexPatterns,
+                            settings: { index: settings },
+                            mappings: { _doc: mappings },
+                            aliases
+                        }
+                    });
+                }
+            } else {
+                expect(resp).toEqual({
+                    [tempName]: {
+                        order: 0,
+                        index_patterns: indexPatterns,
+                        settings: { index: settings },
+                        mappings,
+                        aliases
+                    }
+                });
+            }
+        });
+    });
+
+    describe('indices.deleteTemplate', () => {
+        const tempName = 'test-template-delete';
+
+        beforeAll(async () => {
+            await wrappedClient.indices.putTemplate({
+                name: tempName,
+                body: {
+                    index_patterns: ['test-delete-template'],
+                    settings: {
+                        number_of_shards: 3,
+                        number_of_replicas: 2
+                    },
+                    mappings: {
+                        properties: {
+                            name: { type: 'keyword' },
+                            uuid: { type: 'keyword' },
+                            created: { type: 'date' }
+                        }
+                    },
+                    aliases: {
+                        template_test: {}
+                    }
+                }
+            });
+        });
+
+        it('should delete the template', async () => {
+            const params: ClientParams.IndicesDeleteTemplateParams = {
+                name: tempName,
+                master_timeout: '60s'
+            };
+
+            const resp = await wrappedClient.indices.deleteTemplate(params);
+
+            expect(resp.acknowledged).toBeTrue();
+        });
+    });
+
+    describe('indices.existsTemplate', () => {
+        const tempName = 'test-template-exists';
+
+        beforeAll(async () => {
+            await wrappedClient.indices.putTemplate({
+                name: tempName,
+                body: {
+                    index_patterns: ['test-template-exists*'],
+                    settings: {
+                        number_of_shards: 3,
+                        number_of_replicas: 2
+                    },
+                    mappings: {
+                        properties: {
+                            name: { type: 'keyword' },
+                            uuid: { type: 'keyword' },
+                            created: { type: 'date' }
+                        }
+                    },
+                    aliases: {
+                        template_test: {}
+                    }
+                }
+            });
+        });
+
+        afterAll(async () => {
+            await client.indices.deleteTemplate({ name: tempName });
+        });
+
+        it('should return true if template exists', async () => {
+            const params: ClientParams.IndicesExistsTemplateParams = {
+                name: tempName,
+                master_timeout: '60s',
+                local: false,
+                flat_settings: false
+            };
+
+            const resp = await wrappedClient.indices.existsTemplate(params);
+
+            expect(resp).toBeTrue();
+        });
+
+        it('should return false if template does not exist', async () => {
+            const params: ClientParams.IndicesExistsTemplateParams = {
+                name: 'not-exists',
+                master_timeout: '60s',
+                local: false,
+                flat_settings: false
+            };
+
+            const resp = await wrappedClient.indices.existsTemplate(params);
+
+            expect(resp).toBeFalse();
+        });
+    });
+
+    xdescribe('indices.getIndexTemplate', () => {
+        const tempName = 'test-template-get';
+
+        const settings = {
+            number_of_shards: '3',
+            number_of_replicas: '2'
+        };
+
+        const mappings = {
+            properties: {
+                name: { type: 'keyword' },
+                uuid: { type: 'keyword' },
+                created: { type: 'date' }
+            }
+        };
+
+        const indexPatterns = ['test-template-get*'];
+
+        const aliases = { template_test: {} };
+        let newClient: any;
+
+        beforeAll(async () => {
+            const createdClient = await createClient({ node: host }, testLogger);
+
+            newClient = createdClient.client;
+
+            if (majorVersion !== 6) {
+                await newClient.indices.putIndexTemplate({
+                    name: tempName,
+                    body: {
+                        index_patterns: indexPatterns,
+                        template: {
+                            settings,
+                            mappings,
+                            aliases
+                        }
+                    }
+                });
+            }
+        });
+
+        afterAll(async () => {
+            if (majorVersion !== 6) {
+                await newClient.indices.deleteIndexTemplate({ name: tempName });
+            }
+        });
+
+        it('should return template', async () => {
+            const params = {
+                name: tempName
+            };
+
+            if (majorVersion !== 6) {
+                const resp = await wrappedClient.indices.getIndexTemplate(params);
+
+                expect(resp).toEqual({
+                    index_templates: [
+                        {
+                            name: tempName,
+                            index_template: {
+                                index_patterns: indexPatterns,
+                                template: {
+                                    settings: { index: settings },
+                                    mappings,
+                                    aliases
+                                },
+                                composed_of: []
+                            },
+                        }
+                    ]
+                });
+            }
+        });
+    });
+
+    describe('tasks.list', () => {
+        it('should return tasks', async () => {
+            const params: ClientParams.TasksListParams = {
+                group_by: 'nodes',
+                timeout: '60s'
+            };
+
+            const resp = await wrappedClient.tasks.list(params);
+
+            expect(resp.nodes).toBeDefined();
+        });
+
+        it('should return tasks with more parameters', async () => {
+            const params: ClientParams.TasksListParams = {
+                group_by: 'none',
+                timeout: '60s',
+                detailed: true
+            };
+
+            const resp = await wrappedClient.tasks.list(params);
+
+            expect(resp.tasks).toBeDefined();
         });
     });
 });
